chore: utils function created for reuse
