+++ packages/hoppscotch-common/src/components/http/authorization/OAuth2.vue
@@ -182,11 +182,7 @@
 </template>
 
 <script setup lang="ts">
-import {
-  HoppGQLAuthOAuth2,
-  HoppRESTAuthOAuth2,
-  parseTemplateStringE,
-} from "@hoppscotch/data"
+import { HoppGQLAuthOAuth2, HoppRESTAuthOAuth2 } from "@hoppscotch/data"
 import { useService } from "dioc/vue"
 import * as E from "fp-ts/Either"
 import { Ref, computed, ref } from "vue"
@@ -194,7 +190,7 @@ import { z } from "zod"
 import { useI18n } from "~/composables/i18n"
 import { refWithCallbackOnChange } from "~/composables/ref"
 import { useToast } from "~/composables/toast"
-import { getCombinedEnvVariables } from "~/helpers/preRequest"
+import { replaceTemplateStringsInObjectValues } from "~/helpers/auth"
 import { AggregateEnvironment } from "~/newstore/environments"
 import authCode, {
   AuthCodeOauthFlowParams,
@@ -1055,55 +1051,6 @@ const generateOAuthToken = async () => {
     toast.error(errorMessages[res.left])
     return
   }
-}
-
-const replaceTemplateStringsInObjectValues = <
-  T extends Record<string, unknown>,
->(
-  obj: T
-) => {
-  const envs = getCombinedEnvVariables()
-
-  const requestVariables =
-    props.source === "REST"
-      ? restTabsService.currentActiveTab.value.document.request.requestVariables.map(
-          ({ key, value }) => ({
-            key,
-            value,
-            secret: false,
-          })
-        )
-      : []
-
-  // Ensure request variables are prioritized by removing any selected/global environment variables with the same key
-  const selectedEnvVars = envs.selected.filter(
-    ({ key }) =>
-      !requestVariables.some(({ key: reqVarKey }) => reqVarKey === key)
-  )
-  const globalEnvVars = envs.global.filter(
-    ({ key }) =>
-      !requestVariables.some(({ key: reqVarKey }) => reqVarKey === key)
-  )
-
-  const envVars = [...selectedEnvVars, ...globalEnvVars, ...requestVariables]
-
-  const newObj: Partial<T> = {}
-
-  for (const key in obj) {
-    const val = obj[key]
-
-    if (typeof val === "string") {
-      const parseResult = parseTemplateStringE(val, envVars)
-
-      newObj[key] = E.isRight(parseResult)
-        ? (parseResult.right as T[typeof key])
-        : (val as T[typeof key])
-    } else {
-      newObj[key] = val
-    }
-  }
-
-  return newObj as T
 }
 
 const grantTypeTippyActions = ref<HTMLElement | null>(null)
